pi-requirements.txt
Adafruit-Blinka==1.1.0
adafruit-circuitpython-neopixel==3.3.4
Adafruit-PlatformDetect==0.0.9
Adafruit-PureIO==0.2.3
aiohttp==3.5.4
aiojobs==0.2.2
async-timeout==3.0.1
attrs==18.2.0
backcall==0.1.0
chardet==3.0.4
decorator==4.3.0
idna==2.8
idna-ssl==1.1.0
ipython==7.2.0
ipython-genutils==0.2.0
jedi==0.13.2
multidict==4.5.2
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
prompt-toolkit==2.0.7
ptyprocess==0.6.0
Pygments==2.3.1
rpi-ws281x==4.1.0
RPi.GPIO==0.6.5
six==1.12.0
spidev==3.2
traitlets==4.3.2
typing-extensions==3.7.2
wcwidth==0.1.7
yarl==1.3.0
